move into googlespanner package
