+++ bin/bindings.properties
@@ -42,7 +42,7 @@ elasticsearch:com.yahoo.ycsb.db.ElasticsearchClient
 geode:com.yahoo.ycsb.db.GeodeClient
 googlebigtable:com.yahoo.ycsb.db.GoogleBigtableClient
 googledatastore:com.yahoo.ycsb.db.GoogleDatastoreClient
-googlespanner:com.yahoo.ycsb.db.GoogleSpannerClient
+googlespanner:com.yahoo.ycsb.db.googlespanner.GoogleSpannerClient
 hbase094:com.yahoo.ycsb.db.HBaseClient
 hbase098:com.yahoo.ycsb.db.HBaseClient
 hbase10:com.yahoo.ycsb.db.HBaseClient10
