+++ googlespanner/src/main/java/com/yahoo/ycsb/db/googlespanner/package-info.java
@@ -19,4 +19,4 @@
  * The YCSB binding for Google's <a href="https://cloud.google.com/spanner/">
  * Spanner</a>.
  */
-package com.yahoo.ycsb.db;
+package com.yahoo.ycsb.db.googlespanner;
